pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "r3-mm-pipeline"
version = "0.1.0"
description = "Multiple Myeloma single-cell computational biology pipeline with agentic tuning"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Abhignya Jagathpally", email = "abhignya.j@gmail.com"}
]
keywords = ["bioinformatics", "single-cell", "multiple-myeloma", "scRNA-seq", "agentic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "scanpy>=1.9.6",
    "anndata>=0.9.2",
    "scvi-tools>=1.1.2",
    "harmony-pytorch>=0.1.8",
    "celltypist>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    "zarr>=2.16.0",
    "mlflow>=2.10.0",
    "wandb>=0.16.0",
    "dvc>=3.34.0",
    "ray[tune]>=2.8.0",
    "dask[dataframe]>=2023.12.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "GEOparse>=2.0.3",
    "torch>=2.1.0",
    "lightning>=2.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "umap-learn>=0.5.3",
    "leidenalg>=0.10.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "pre-commit>=3.3.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/abhignya-j/r3-mm-pipeline"
Repository = "https://github.com/abhignya-j/r3-mm-pipeline.git"
Issues = "https://github.com/abhignya-j/r3-mm-pipeline/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
r3_mm_pipeline = ["configs/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
select = [
    "E",
    "F",
    "W",
    "C",
    "I",
    "N",
]
ignore = [
    "E501",  # Line too long
    "E203",  # Whitespace before ':'
]
target-version = "py311"

[tool.ruff.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
